src/requirements.txt
httpx==0.28.1
pynacl==1.5.0